move undefined type var to resolver
